requirement.txt
python-docx==0.8.11
langid==1.1.6